+++ src/app/poll/poll-selection/poll-selection.component.html
@@ -1,0 +1,12 @@
+<form [formGroup]="pinForm" (submit)="confirmPin()" *ngIf="!polls">
+    <input formControlName="pin">
+    <button class="btn btn-primary" type="submit">{{'show-event.continue'|translate}}</button>
+</form>
+
+<div *ngIf="polls && polls.length > 0">
+    <ul>
+        <li *ngFor="let poll of polls">
+            <a [routerLink]="[poll.id]" queryParamsHandling="merge">{{poll.title[translate.currentLang]}}</a>
+        </li>
+    </ul>
+</div>
