+++ src/app/poll/poll-selection/poll-selection.component.ts
@@ -1,0 +1,55 @@
+import { Component, OnInit } from '@angular/core';
+import { FormGroup, FormBuilder } from '@angular/forms';
+import { Poll } from '../model/poll';
+import { ActivatedRoute, Router } from '@angular/router';
+import { PollService } from '../shared/poll.service';
+import { TranslateService } from '@ngx-translate/core';
+import { combineLatest } from 'rxjs';
+
+@Component({
+  selector: 'app-poll-selection',
+  templateUrl: './poll-selection.component.html',
+  styleUrls: ['./poll-selection.component.scss']
+})
+export class PollSelectionComponent implements OnInit {
+
+  pinForm: FormGroup;
+  polls: Poll[];
+  eventShortName: string;
+
+  constructor(
+    private route: ActivatedRoute, 
+    private router: Router,
+    private pollService: PollService, 
+    public translate: TranslateService,
+    private fb: FormBuilder) { }
+
+  ngOnInit(): void {
+    this.pinForm = this.fb.group({pin: null});
+    combineLatest([this.route.params, this.route.queryParams]).subscribe(([params, query]) => {
+      this.eventShortName = params['eventShortName'];
+      console.log(query);
+      if (query['pin']) {
+        this.loadPolls(this.eventShortName, query['pin']);
+      }
+    });
+  }
+
+  private loadPolls(eventShortName: string, pin: string) {
+    this.pollService.getAllPolls(eventShortName, pin).subscribe(res => {
+      if (res.success) {
+        this.router.navigate([], {queryParamsHandling: 'merge', queryParams: {pin: pin}});
+        this.polls = res.value;
+        if (this.polls.length === 1) {
+          this.router.navigate([this.polls[0].id], {relativeTo: this.route, queryParamsHandling: 'merge', queryParams: {pin: pin}});
+        }
+      }
+    });
+  }
+
+  confirmPin() {
+    const pin = this.pinForm.value.pin;
+    this.loadPolls(this.eventShortName, pin);
+  }
+
+}
